refactor: move dict-related functions to a more suitable place
